make default font a skosh bigger
this took a while to figure out. Info said that setting the default
font via the Options menu and then using save options should write it
out and save it, but I couldn't find where it wrote it, and neither
could a subsequent start of Emacs.

So I did the set default font, then looked at the font with C-u C-x =
and copied the listed font out and used that string in the init.el
file.

Also:

- pick up a new ruby snippet to make my life easier :)